// src/pages/WMS/types/activitysubgroup-wms.ts
export type TActivitysubgroup = {
    activity_subgroup_code: string;
    act_subgroup_name: string;
    company_code?: string;
    mandatory_flag?: string;
    act_group_code?: string;
    validate_flag?:string;
    updated_at?: Date;
    updated_by?: string;
    created_by?: string;
    created_at?: Date;
  };